created file for the test of the question function actions
